requirements.txt
async-timeout==4.0.2
black==22.8.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Deprecated==1.2.13
idna==3.4
mypy-extensions==0.4.3
numpy==1.23.3
packaging==21.3
pandas==1.4.4
pathspec==0.10.1
platformdirs==2.5.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
six==1.16.0
spotipy==2.20.0
tomli==2.0.1
urllib3==1.26.12
wrapt==1.14.1
